Retweet text through store, add CreatePost component
